+++ native/scraper.bat
@@ -1,5 +1,4 @@
 @echo off
 ::pip install selenium
 ::pip install beautifulsoup4
-call scraper.py
-python -u txt-parser.py
+python -u native-host.py
