+++ options.html
@@ -7,7 +7,8 @@
     <!--Add in some options to change-->
     <input type="text" id="link-input">
     <button id="save">Save</button>
-    <h1></h1>
+    <button id="run">Run</button>
+    <h1>Output goes here</h1>
 
     <script src="options.js"></script>
   </body>
