+++ popup.html
@@ -2,10 +2,8 @@
 <!DOCTYPE html>
 <html>
   <body>
-    <button class="button" id="start">
-      Run
-    </button>
+    <a href="chrome-extension://hffkpdiheoifjamngjellaojlbdmbejh/options.html" target="_blank">Options</a>
     <p id="output"></p>
-    <script src="popup.js"></script>
+    <!-- <script src="popup.js"></script> -->
   </body>
 </html>
